refactor(number-field): revamped number field component
