+++ package.json
@@ -72,6 +72,7 @@
         "test:ci": "yarn test:start",
         "test:create": "wireit",
         "test:errors": "yarn test | grep -A 32 ❌",
+        "test:file": "npx web-test-runner --verbose --config ./web-test-runner.single-file.config.js",
         "test:focus": "yarn build && yarn test:ci --group",
         "test:start": "web-test-runner",
         "test:visual": "yarn test:visual:ci",
